Redesign Skeleton, create own class files and meaningful paths
