+++ 01_Exercise_Supermarket/Simulation.py
@@ -1,9 +1,14 @@
+
 from collections import deque
 import heapq
+import os
+from classes import Customer, Event, EventQueue, Station
 
-f = open("supermarkt.txt", "w")
-fc = open("supermarkt_customer.txt", "w")
-fs = open("supermarkt_station.txt", "w")
+#-------------------------------------------------------------------------------------------------- File handling
+directory_path = os.getcwd()
+f = open(directory_path + "\\01_Exercise_Supermarket\data\supermarkt.txt", "w")
+fc = open(directory_path + "\\01_Exercise_Supermarket\data\supermarkt_customer.txt", "w")
+fs = open(directory_path + "\\01_Exercise_Supermarket\data\supermarkt_station.txt", "w")
 
 
 # print on console and into supermarket log
@@ -16,7 +21,7 @@ def my_print(msg):
 # k: customer name
 # s: station name
 def my_print1(k, s, msg):
-    t = EvQueue.time
+    t = EventQueue.time
     print(str(round(t, 4)) + ':' + k + ' ' + msg + ' at ' + s)
     fc.write(str(round(t, 4)) + ':' + k + ' ' + msg + ' at ' + s + '\n')
 
@@ -25,58 +30,9 @@ def my_print1(k, s, msg):
 # s: station name
 # name: customer name
 def my_print2(s, msg, name):
-    t = EvQueue.time
+    t = EventQueue.time
     # print(str(round(t,4))+':'+s+' '+msg)
     fs.write(str(round(t, 4)) + ':' + s + ' ' + msg + ' ' + name + '\n')
-
-
-# class consists of instance variables:
-# t: time stamp
-# work: job to be done
-# args: list of arguments for job to be done
-# prio: used to give leaving, being served, and arrival different priorities
-class Ev:
-    counter = 0
-
-    def __init__(self, t, work, args=(), prio=255):
-        self.t = t
-        self.n = Ev.counter
-        self.work = work
-        self.args = args
-        self.prio = prio
-        Ev.counter += 1
-
-
-# class consists of
-# q: event queue
-# time: current time
-# evCount: counter of all popped events
-# methods push, pop, and start as described in the problem description
-
-class EvQueue:
-# please implement here
-
-
-# class consists of
-# name: station name
-# buffer: customer queue
-# delay_per_item: service time
-# CustomerWaiting, busy: possible states of this station
-class Station():
-# please implement here
-
-
-# class consists of
-# statistics variables
-# and methods as described in the problem description
-class Customer():
-    served = dict()
-    dropped = dict()
-    complete = 0
-    duration = 0
-    duration_cond_complete = 0
-    count = 0
-# please implement here
 
 
 
@@ -85,13 +41,13 @@ def startCustomers(einkaufsliste, name, sT, dT, mT):
     t = sT
     while t < mT:
         kunde = Customer(list(einkaufsliste), name + str(i), t)
-        ev = Ev(t, kunde.run, prio=1)
+        ev = Event(t, kunde.run, prio=1)
         evQ.push(ev)
         i += 1
         t += dT
 
 
-evQ = EvQueue()
+evQ = EventQueue()
 baecker = Station(10, 'Bäcker')
 metzger = Station(30, 'Metzger')
 kaese = Station(60, 'Käse')
@@ -109,7 +65,7 @@ einkaufsliste2 = [(30, metzger, 2, 5), (30, kasse, 3, 20), (20, baecker, 3, 20)]
 startCustomers(einkaufsliste1, 'A', 0, 200, 30 * 60 + 1)
 startCustomers(einkaufsliste2, 'B', 1, 60, 30 * 60 + 1)
 evQ.start()
-my_print('Simulationsende: %is' % EvQueue.time)
+my_print('Simulationsende: %is' % EventQueue.time)
 my_print('Anzahl Kunden: %i' % (Customer.count
                                 ))
 my_print('Anzahl vollständige Einkäufe %i' % Customer.complete)
