+++ 01_Exercise_Supermarket/classes/Customer.py
@@ -1,0 +1,12 @@
+# class consists of
+# statistics variables
+# and methods as described in the problem description
+class Customer():
+    served = dict()
+    dropped = dict()
+    complete = 0
+    duration = 0
+    duration_cond_complete = 0
+    count = 0
+    
+# please implement here
