+++ 01_Exercise_Supermarket/classes/EventQueue.py
@@ -1,0 +1,11 @@
+# class consists of
+# q: event queue
+# time: current time
+# evCount: counter of all popped events
+# methods push, pop, and start as described in the problem description
+
+class EventQueue:
+  
+  def __init__(self):
+    self.removeThisVar = 0
+# please implement here
