+++ 01_Exercise_Supermarket/classes/Station.py
@@ -1,0 +1,10 @@
+# class consists of
+# name: station name
+# buffer: customer queue
+# delay_per_item: service time
+# CustomerWaiting, busy: possible states of this station
+class Station():
+  
+  def __init__(self):
+    self.removeThisVar = 0
+  # please implement here
